add component Select, which according to prop 'multiSelect' renders appropriate component
