+++ index.js
@@ -1,4 +1,5 @@
 export { default as Grid } from "./lib/Grid";
 export { default as CodeInput } from "./lib/CodeInput";
-export { default as Select } from "./lib/Select";
+export { default as SingleSelect } from "./lib/SingleSelect";
 export { default as MultiSelect } from "./lib/MultiSelect";
+export { default as Select } from './lib/Select';
